+++ internal/corpus/programs/rolling_average.c
@@ -6,18 +6,18 @@
 //
 // Pressure is fractional, so the samples and both aggregates are doubles and
 // the mean is a plain division rather than a truncating one. The cursor stays
-// a long long: it counts slots.
+// a long: it counts slots.
 
 const dev gauge = d0;
 const dev display = d1;
 const dev alarm = d2;
 
-constexpr long long kWindow = 8;
+constexpr long kWindow = 8;
 constexpr double kAlarmKilopascals = 200.5;
 
 double samples[kWindow];
-long long cursor;
-long long filled;
+long cursor;
+long filled;
 
 void record(double value) {
     samples[cursor] = value;
@@ -33,7 +33,7 @@ double average(void) {
     }
 
     double sum = 0.0;
-    for (long long i = 0; i < filled; i++) {
+    for (long i = 0; i < filled; i++) {
         sum += samples[i];
     }
     return sum / filled;
@@ -41,7 +41,7 @@ double average(void) {
 
 double peak(void) {
     double best = 0.0;
-    long long i;
+    long i;
     for (i = 0; i < filled; i++) {
         if (samples[i] > best) {
             best = samples[i];
